+++ src/main/java/logan/pickpocket/main/PickpocketPlugin.java
@@ -18,6 +18,7 @@ import logan.pickpocket.commands.DebugGiveRandom;
 import logan.pickpocket.commands.MainCommand;
 import logan.pickpocket.commands.ReloadCommand;
 import logan.pickpocket.commands.SetSkillCommand;
+import logan.pickpocket.commands.SkillsCommand;
 import logan.pickpocket.commands.StatusCommand;
 import logan.pickpocket.commands.TestCommand;
 import logan.pickpocket.commands.ToggleCommand;
@@ -111,6 +112,7 @@ public class PickpocketPlugin extends JavaPlugin {
         CommandDispatcher.registerCommand(new DebugGiveRandom());
         CommandDispatcher.registerCommand(new ReloadCommand());
         CommandDispatcher.registerCommand(new SetSkillCommand());
+        CommandDispatcher.registerCommand(new SkillsCommand());
         CommandDispatcher.registerCommand(new ToggleCommand());
         CommandDispatcher.registerCommand(new StatusCommand());
         CommandDispatcher.registerCommand(new TestCommand());